requirements.txt
streamlit
llama-index
llama-index-llms-groq
llama-index-embeddings-huggingface
chromadb
python-dotenv
pypdf
sentence-transformers
